Move FullScreenButton to MenuBar directory
