+++ src/components/MenuBar/MenuBar.tsx
@@ -6,7 +6,7 @@ import LocalAudioLevelIndicator from './LocalAudioLevelIndicator/LocalAudioLevel
 import Button from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import TextField from '@material-ui/core/TextField';
-import ToggleFullscreenButton from '../ToggleFullScreenButton/ToggleFullScreenButton';
+import ToggleFullscreenButton from './ToggleFullScreenButton/ToggleFullScreenButton';
 import Toolbar from '@material-ui/core/Toolbar';
 import Menu from './Menu/Menu';
